Restructure Tile Engine's benchmarking process
This change restructures the Benchmark structs into 3 files.
There is an addition of a base class for all GEMM benchmarks, derived classes for
Universal GEMM, multi dim GEMM, and GEMM preshuffle. Common functions have been relocated
into a common directory. For any derived base classes, only the redefination of the
constructor is needed, significantly mitigating the need for duplicated code.

Restructure Tile Engine's profiling process

This change restructures the profiling process in Tile Engine into
a base class for the Profiling and Problem structs. With this all files
needed for Tile Engine will have a base struct and files in the gemm/
directory that can be extended for each GEMM variant. Only the Problem
and Profiler structs along with the reference functions need to be
defined. Profiling functions that are common to each operation have
been moved into a common utility file.

Adding README back into the gemm directory and integrate new preshuffle functions

disabling the gemm tile engine tests and updating preshuffle example to match new tensor_shuffle_utils interface
